Refactor Testforms to have much less repetition
