+++ app/frontend/js/TestEnrichmentExtraction.js
@@ -1,44 +1,27 @@
 
-import { request } from "./utils/request";
+import { bindTestForm } from "./utils/TestForm";
 import { EditorState } from "@codemirror/state";
 import { EditorView, basicSetup } from "codemirror";
 import { json } from "@codemirror/lang-json";
 
-document.addEventListener(
-  "DOMContentLoaded",
-  function () {
-    function sendData(form) {
-      // Bind the FormData object and the form element
-      const FD = new FormData(form);
-      FD.set("_method", "post");
+bindTestForm('test_enrichment_extraction', 'js-test-enrichment-extraction', (response, _alertClass) => {
+  let editor = new EditorView({
+    state: EditorState.create({
+      extensions: [basicSetup, json(), EditorState.readOnly.of(true)],
+      doc: JSON.stringify(JSON.parse(response.data.body), null, 2),
+    }),
+    parent: document.body,
+  });
 
-      const path = `${form.action.replace(/\/\d+$/, "")}/test_enrichment_extraction`;
-      // Fetches the response
-      
-      request.post(path, FD).then(function (response) {
-        let record_selector_viewer_editor = new EditorView({
-          state: EditorState.create({
-            extensions: [basicSetup, json(), EditorState.readOnly.of(true)],
-            doc: JSON.stringify(JSON.parse(response.data.body), null, 2),
-          }),
-          parent: document.body,
-        });
-
-        document.querySelector("#js-enrichment-extraction-result").innerHTML = "";
-        document
-          .querySelector("#js-enrichment-extraction-result")
-          .append(record_selector_viewer_editor.dom);
-      });
-    }
-
-    const button = this.getElementById("js-test-enrichment-extraction");
-
-    if (button) {
-      const form = button.closest("form");
-      button.addEventListener("click", (event) => {
-        sendData(form);
-      });
-    }
-  },
-  false
-);
+  document.querySelector("#js-enrichment-extraction-result").innerHTML = "";
+  document
+    .querySelector("#js-enrichment-extraction-result")
+    .append(editor.dom);
+}, () => {
+  document.querySelector("#js-enrichment-extraction-result").innerHTML = "";
+  document.getElementById(
+    "js-enrichment-extraction-result"
+  ).innerHTML = `<div class="alert alert-danger my-2" role="alert">
+    Something went wrong fetching your enrichment from the Enrichment URL. Please confirm that your Enrichment URL is correct.
+  </div>`;
+});
