+++ app/frontend/js/TestTransformationRecordSelector.js
@@ -1,0 +1,19 @@
+import { bindTestForm } from './utils/TestForm';
+import { EditorState } from "@codemirror/state";
+import { EditorView, basicSetup } from "codemirror";
+import { json } from "@codemirror/lang-json";
+
+bindTestForm('test', 'js-test-transformation-record-selector', (response, _alertClass) => {
+    let editor = new EditorView({
+      state: EditorState.create({
+        extensions: [basicSetup, json(), EditorState.readOnly.of(true)],
+        doc: JSON.stringify(response.data, null, 2),
+      }),
+      parent: document.body,
+    });
+
+    document.querySelector("#js-record-selector-result").innerHTML = "";
+    document
+      .querySelector("#js-record-selector-result")
+      .append(editor.dom);
+});
