+++ app/views/transformation_definitions/_form.html.erb
@@ -17,7 +17,7 @@
   </div>
 
   <div class="col-12">
-    <button type="button" class="btn btn-secondary me-1" id="js-test-transformation">
+    <button type="button" class="btn btn-secondary me-1" id="js-test-transformation-record-selector">
       Test Record Selector
     </button>
